Move shot sampling to measurement objects
